Refactor generate dart script
